Update lcd.vhd
This code displays 0123456 when you keep pressing reset button. That is when reset is pressed the code runs otherwise it is paused. On each press of the reset button as expected everything is reseted. 

In Port - start, lcd_write, input are to be left blank - i.e. it will be used when another higher level entity is to be defined.